requirements.txt
langchain>=0.2.10
langchain-openai>=0.1.11
openai>=1.40.0
python-dotenv>=1.0.1
pandas>=2.2.2
matplotlib>=3.8.4
dateparser>=1.2.0
pydantic>=2.8.2
